requirements.txt
Flask==2.3.3
numpy==1.25.2
Pillow==10.0.0
python-dotenv==1.0.0
scikit_learn==1.3.0
torch==2.0.1
torchvision==0.15.2
webcolors==1.13
